Make Heatmaps for Co-occurring mutations

Create heatmaps using the ComplexHeatmap package (one version with all mutated sequences and one version with sequences with high-risk mutations); heatmaps are colored according to whether an assay is affected by mutated sequences

makeCoOccHeatmaps(
  DB,
  startDate = NULL,
  endDate = NULL,
  archiveDBFile = NULL,
  saveTable = FALSE,
  saveFolder = NULL,
  all_width = NULL,
  all_height = NULL,
  hr_width = NULL,
  hr_height = NULL,
  counts_font_size = NULL
)

Arguments

DB

File path to subset database file. In the case of making summary tables, we use the 3M version from CoMIT, returned from pull_subset_DB()

startDate

start date for subset DB, if NULL then the .1 percent analysis won't be run

endDate

end date for subset DB, if NULL then the .1 percent analysis won't be run

archiveDBFile

File path to archive db (used to retrieve older subset of data for comparison; older subset uses same length of time as input subset db, e.g. start_date is "2022-5" and end_date is "2022-07", then older subset start date is "2022-02 and "2022-04)

saveTable

logical value indicating whether to save the tables to the specified saveFolder; # NOTE: It is easiest to always set saveTable to TRUE. If it is set to FALSE, there is a chance the legend will be cut off, and the user will have to adjust the proportions in the Plots pane for each table before group_block_anno is run. There is a pause in execution added to the code right before the subtitle is added in order to give the user time to do this.

saveFolder

file path containing folder the finished tables will be saved in

all_width

width of cells for heatmap showing combinations for all mutations

all_height

height of cells for heatmap showing combinations for all mutations

hr_width

width of cells for heatmap showing combinations for high-risk mutations

hr_height

height of cells for heatmap showing combinations for high-risk mutations

counts_font_size

font size for counts at the bottom of the bar chart