Make Dataframe for Co-occurring Mutations Heatmap

Create dataframe to be used by makeCoOccHeatmaps() in creating visulization of co-occurring mutations on different combinations of assays, with the sequences organized according to WHO designations/Pango lineages

makeDFcoOccHeatmap(
  DB,
  HR = FALSE,
  BASE_PAIR_LIMIT = 10,
  threeAssay = FALSE,
  startDate = NULL,
  endDate = NULL,
  archiveDBFile = NULL
)

Arguments

DB

File path to subset database file. In the case of making summary tables, we use the 3M version from CoMIT, returned from pull_subset_DB()

HR

if TRUE, returns 3 prime end, i.e. high risk, version

BASE_PAIR_LIMIT

number of base pairs from the 3 prime end to be considered high risk

threeAssay

if TRUE, makes three-assay version of dataframe (a, d, and e assays) for 3-assay version of Covid test

startDate

start date for subset DB, if NULL then the .1 percent analysis won't be run

endDate

end date for subset DB, if NULL then the .1 percent analysis won't be run

archiveDBFile

File path to archive db (used to retrieve older subset of data for comparison; older subset uses same length of time as input subset db, e.g. start_date is "2022-5" and end_date is "2022-07", then older subset start date is "2022-02 and "2022-04)

Value

comboFreq, a dataframe with counts and percentages of each co-occurring mutation combo of assays