Make Database Breakdown Dataframe V2

Make Database Breakdown Dataframe V2

makeDFdatabaseBreakdown_2(DB, interval, lastMonth, total_seqs, total_seqs_1m)

Arguments

DB

File path to subset database file. In the case of making summary tables, we use the 3M version from CoMIT, returned from pull_subset_DB()

interval

string describing the first period analyzed, which is a 3-month period, e.g. May - July 2022

lastMonth

string describing the second period analyzed, which is a 1-month period, e.g. July 2022

total_seqs

total number of sequences for the 3-month period

total_seqs_1m

total number of sequences for the 1-month period

Value

combined_df, a dataframe with data for both time periods and all needed formatting applied for creation of the gt table