Make Primer Variant Mutation Dataframe

Make Primer Variant Mutation Dataframe

makePrimerVarMutationDF(
  start_date,
  end_date,
  varIds = NULL,
  archiveDBFile,
  subsetDBFile = NULL,
  printableDoubles = FALSE
)

Arguments

start_date

Start date for subset db

end_date

End date for subset db

varIds

Vector of variant IDs to be used in the visualization, provided if subsetDBFile is NULL (not using Sig_Var_IDs from the subsetDBFile)

archiveDBFile

File path to archive db (used to retrieve older subset of data for comparison; older subset uses same length of time as input subset db, e.g. start_date is "2022-5" and end_date is "2022-07", then older subset start date is "2022-02 and "2022-04)

subsetDBFile

File path to subset db, provided if using variant IDs from Sig_Var_IDs as those in the visualization (varIds parameter should be NULL if this is provided)

printableDoubles

Boolean, default is FALSE, if TRUE will show double mutations for a single variant ID in a single column of the visualization

Value

unique_var_table, a table with mutations and trends for each significant variant, comparing this subset to an older subset