Make Database Breakdown Table

Make a table showing sequence counts categorized by lineage, comparing a given 3-month period to the last month of the 3-month period

makedbBreakdownTable(
  DB,
  variantFile,
  interval,
  lastMonth,
  colWidth,
  extraNote = NULL,
  saveTable = FALSE,
  saveFolder = NULL,
  height = NULL,
  width = NULL
)

Arguments

DB

File path to subset database file. In the case of making summary tables, we use the 3M version from CoMIT, returned from pull_subset_DB()

variantFile

csv file containing "Displayed_Lineage_Order" column for ordering row names/row groups in the GT table

interval

string describing the first period analyzed, which is a 3-month period, e.g. May - July 2022

lastMonth

string describing the second period analyzed, which is a 1-month period, e.g. July 2022

colWidth

desired column width; increase if text is wrapping (applies to all columns, for both printed and saved version)

extraNote

string with extra source note to be displayed at the bottom of the table

saveTable

logical value indicating whether to save the tables to the specified saveFolder

saveFolder

file path containing folder the finished tables will be saved in

height

desired height of table .png images (when saveTable is TRUE)

width

desired width of table .png images (when saveTable is TRUE)