Update DB (Add sequence info to a database)

CoMIT Database update. Requires an existing CoMIT database. Pulls from the CSV specified

update_db(
  dbName,
  dbFile,
  saveCopy,
  pullDate_1,
  pullDate_2,
  fasta_list_file,
  run_histories_filepath
)

Arguments

dbName

string name of the database, with no filepath or .db extension (e.g. "All_Assay_Archive")

dbFile

string file path of the database, with .db extension

saveCopy

logical value; if true, will save a copy of the old DB called dbName to _DB_files/DB_Copies in the user's home directory

pullDate_1

string date for beginning of date selection used in the update (refers to pullDate column in fasta_list_file, only things >= this date will be included in comit_classify run), e.g. "2022-10-11"

pullDate_2

string date for end of date selection used in the update (refers to pullDate column in fasta_list_file, only things <= this date will be included in comit_classify run); if NA, use everything after pullDate_1

fasta_list_file

string name of the csv file with fasta file information, including path to file; has columns folder (location of each fasta file), fastaFile (fasta file name), pullDate, metaDataFile (NA if same name as fasta file, name if different)

run_histories_filepath

string path to .txt file used to record run histories for this database (file will be called run_histories_file + db name + .txt)